Added nav bar and basic title page text
Added a nav bar that will eventually link to different sections of the page. Made it responsive by shrinking nav bar to a bar icon that will eventually become a drop down nav bar (not working yet). Added bare text to title page. Centered and div is formatted as desired.